Factorize function 'encrypt_one_commit'
By factorizing 'encrypt_one_commit' into parts, I can delay the
archiving and encrypting to a later stage, do it for all files
of the push, instead of a commit, to minimize the resulting file
size, because we compress all files together onece in gpg. And
create only one encrypted commit for a push.

 - copy_out_objects for copying objects of a commit
 - tar_to_stdout for archiving a directory to pipe
 - encrypt_pipe for encrypting data from pipe to pipe
 - pipe_to_object for creating blob from pipe
 - cleanup for cleaning up the temporary files