Add function transform
Copy all object files of the commits to be pushed to a directory,
pack that directory as a whole and encrypt it, and write the
encrypted data to a hash object, then create a tree object for
the newly created hash object, and then a commit object for the
newly created tree object. Commit message is of static format,
it must have a title of 'CIPHER COMMIT' and contain the start and
stop commits' id in the message body. The newest plaintext commit
and newest cipher commit objects' id is saved for later use, the
plaintext form key also keep for later reference.