Add git-shadow-pubkey
Public keys of the shadow git system are for encrypting
the symmetric keys which are used to encrypt the actual
git objects. For every shadow push, all the git objects
to push are encrypted with a new symmetric key, this
symmetric key will then be encrypted with all the public
keys in the pubkeys database of the shadow git, then the
encrypted git objects and the encrypted symmetric key be
pushed to the remote. These public keys normally are of
the team's members.