Correct the object finding problem
When do it in the old way, many objects that not
created by the commit will also be collectd, thus
produces bigger encrypted file, takes much longer
time.

With 'git show' instead of 'git log', files changed
from the view of the commit will be detected, even
if these changes are not introduced by the commit.